requirements.txt
playwright>=1.40.0
anthropic>=0.18.0
pillow>=10.0.0
rich>=13.0.0
